improve startup time with ConsumeMany()
new sarama or kafka are serializing joining consumer groups, and
for consumers which want to join many topics this is beginning to
take a long time. It seems we can't serially join more than one
topic per 4 or 5 seconds.

So this adds a new API Client.ConsumeMany(topics[]string)
which is logically equivalent to calling Client.Consume(topic string)
over and over, but about as fast as a single call to Consume()