arrange things so con.Done() can write to the partition's chan
directly. This makes the code more complicated, but avoids a hop
through a single channel to a single goroutine, neither of which
are good for performance.

(of course the whole sarama idea of returning each kafka msg as
a single chan message is also bad for performance. they should
return an array of such messages. But that would require a forked
sarama)