fix bug: committed offset would seem to jump back a little when
a new message arrived and wasn't finished being processed, and
then jump back forward when that message was Done().

The cause was the adjustment for the partially completed bucket
of offsets was added to the committed offset only when it was
clear what needed to be added. Otherwise, when one or more msgs
were outstanding from the bucket, nothing cound be concluded,
and nothing was added. Thus if at time T1 the 0th bucket has
equal read and done offsets, then the committed offset is the
bucket's base offset + the # done, and then at T2 a message
has been read but isn't Done(), then the committed offset was
just the base offset, and thus would appear to jump back by
a handful (up to 63) offsets.