support for not tracking all received message offsets, and then
using Done() to only occasionaly commit an offset. This allows
callers to call Done(msg) once and mark all the messages in the
same partition with offsets <= msg.Offset as complete.

This is enabled by clearing the new Config.AllowOutOfOlderDone bool
before creating the consumer.

The default is AllowOutOfOlderDone=true in order to remain backwards
compatible.